Fix sequence bit placement.
It really was all about that damn data sequence
bit! The last remaining bug (that I'm aware of
right now anyway) in the bulk data sequence bit
handling was that I was stuffing it in some
reserved bits of the TD list header control byte
rather than the correct DToggle bit. Fixed that
and now we're good to go! All transactions work
even on the crummy HP drive and without throwing
in a million resets.