Cleanup verbose printf logging
Move most of the printf output logging misc. data
during the USB initialization and transfer
operations over to the new LOGV() macro which is
a no-op unless the "verbose" variable is greater
than zero.

This caused issues with the HP drive, which was
NAKing its set address control command right
after reset if the verbose printfs were removed
from run(). This was worked around by putting a
stall in after the reset to give the drive time to
get itself ready, but there must be a way to wait
for the drive deterministically. Will need to look
into this further later.

Also started working on making the code work on
ports != 0. So far this is only plumbed to the
USB reset routine, and is not tested.