Add TDList data structure and helper code
The TDList structure is a slightly-unpacked
version of the EZHost HW TD List data structure.
This data structure is private to the USB driver
code, which now has a bunch of convenience macros
and helper functions to initialize it, pack+write
it to the EZHost memory, and read it back+unpack
it from the EZHost memory to check execution
status. These helpers replace some very difficult
to read hard-coded stuff in the bulkcmd() function
so far, but can also replace the similarly-
difficult-to-read code in the control message
function, and are the first step towards cleanly
enqueing more than one TD per frame in the bulk
command function.

Although the command is considered private to the
driver code, it is declared in the driver public
header to avoid the need to dynamically allocate
it in the future.