Add usbverif.c, a host USB verification program
usbverif.c builds into usbverif.cof, which when
run, attempts to initialize bulk storage devices
on both host USB ports on the Skunkboard, write
2MB of pre-computed random data to each port
respectively from the files rand0.bin and
rand1.bin on the JCP host, then read back 2MB of
data from each port respectively to the files
usbdump0.bin and usbdump1.bin on the JCP host,
then closes the JCP console.

It is expected a script on the JCP host will take
care of ensuring the source files are correctly
generated, and verifying the data read back from
the USB devices on the Skunkboard matches that
in the source files when the program exits.