Add EZHeap data structure to manage EZHost RAM
The EZHost processor has 64k of internal RAM. This
memory is used to store the BIOS/interrupt
routines, the JCP accelerated transfer code, the
JCP <-> skunk communications buffers, and a region
used by the GDB debugger stub. USB TD list and
transfer data must be located in any leftover
memory. After the above allocations, there are
3844 bytes remaining, split over two regions.

Prior to this change, bits of that memory were
statically assigned to the various USB transfer
operations supported by the USB driver code. This
would not scale well when trying to queue up an
arbitrary number of USB transfers rather than one
at a time, so this change introduces a set of heap
management functions to allocate and free memory
from the available free regions.

The bulkcmd() function has been converted over to
use the EZHost heap manager. It still does not
attempt to run more than one transfer at a time,
but this is another baby step in the direction of
queuing up multiple transfers