Add "flash" and "quit" commands to usbffs
Type flash <filename> in the skunk console prompt
to flash a ROM file (<=4MB) to the first bank of
the skunkboard. Don't escape spaces or anything.
The command line processor is incredibly primitive
and interprets anything after the word "flash" and
the first space character following it to be the
filename.

After you see "Flashing complete" on the console,
type "quit" to quit the console, reset the board
with "jcp -r" (Or just power cycle the Jaguar),
then press Up on the D-pad to launch the ROM.