Launch ROMs after flashing them.
Adds some code to reset the EZHost chip and launch
the ROM that was just flashed. This is largely
copied from the similar skunk BIOS code, but it
also attempts to shut down the fancy strobing logo
before launching the ROM.

This works for about half the ROMs I've tested so
far. Others just hang. Need to debug what's wrong
in those cases.