Improve ROM compatibility and support 2MB erases
Set the vertical interrupt register to 0xFFFF
before launching the ROM, matching the setting
from the SkunkBoard BIOS. This fixes the handful
of games that wouldn't launch after flashing
before, though it breaks the original Raiden ROM
just like the current BIOS code does. I'll likely
have to add a work-around specifically for this
ROM at some point.

The code was also modified to erase only the first
2MB of flash for ROMs <= 2MB in size, just like
the default behavior in JCP, which speeds up the
erase step for small ROMS.

As a bonus, a "launch" command was added that
launches the game currently in the first flash
bank.