Scale the logo in/out of the top right corner
More experimentation with the object processor.
Adds code to scale the logo and adjust its X
position such that it shrinks into and back out of
the upper-right corner of the screen.

This involved converting the object list entry to
a scaled bitmap object and adding a bunch of math
to perform the scaling in the list update routine.
A command line handler, "togglescale," was also
added to enable/disable the scaling. It is
disabled by default.