Shut down the GPU before launching ROMs
Various games don't seem to deal well with the GPU
running when they start. Add code to stop the GPU,
and move the code that parks the OLP on the skunk
BIOS's tiny stop object object list to the GPU
shutdown handling code. This seems to get all the
problematic games running directly from a flash
again.