Make ls command list directory on-screen too
Finally moving towards an actual UI, the ls
command now renders the directory listing to the
games list bitmap on screen. Because the string
drawing function is ridiculously slow right now,
also stop listing  the directory on startup. The
games list is not shown by default, so it just
looked like the program was hung if there were
more than one or two directory entries.