Add a test harness for the GPU/GFX/UI code
testgfx.c shows the game list, then, in a loop,
clears it, blits a bunch of text onto it, and
inverts regions of it. The text includes some
corner cases such as unprintable characters and
empty strings. The unprintable character strings
exposed a bug in the GPU code's handling of them,
which is fixed here as well.