Add initial DSP code to poll joystick state
This gets the DSP up and running, and adds basic
parsing of the JOYSTICK and JOYBUTS registers to
extract the state of two standard jaguar
controllers connected to the two built-in
controller ports. Some test code is added to
usbffs.c to dump the decoded button state as well.
Turning the DSP on probably breaks game launching
to some extent as there's currently no effort to
shut the DSP down before launching a game. Other
TODOS:

-Convert diffs in joystick state into timestamped
 events.

-Detect controllers connected via a Team Tap and
 parse their state as well.

-Handle non-standard controller types? Probably
 not worth it for this project, but the joystick
 state extracting code could be useful elsewhere
 as well.