Add button debouncing to input event code
Store a timeout of ~10ms whenever a button state
change is detected, and ignore further state
changes on that button until the timeout is
reached. 10ms was chosen empirically based on
testing on a standard jaguar controller. It caps
the number of state changes per button at 100 per
second and removed all false events except a few
on the option button.