Fix the DSP code that turns the DSP off
The semaphore value used to signal the DSP main
loop to exit and shut down the processor was
using a dword-sized load on a word-sized variable.
While this usually worked out OK, it could
obviously go badly if neighboring memory was
tweaked at all. For reasons I still don't fully
understand, slightly tweaking the offset of the
variable and/or instructions accessing it would
cause this code to hang the DSP, even though the
variable itself is explicitly aligned and the
code segment is phrase aligned.