Handle joypad events
This rather large set of changes finally makes
usbffs useful. It can now be driven entirely
from the Joypad. It can also now be built without
skunklib console support to allow it to run
untethered from the PC, finally achieving stand-
alone ROM flashing support. When built with
skunklib support, it can be controlled from the
skunk console command-line interface, the
joypad, or a combination of both.

Current joypad controls:

Up/Down - Move the file selection around
Option - Refresh the directory listing
Pause - Hide/Show the directory listing
B - Flash the selected ROM and launch it, or
    change to the selected directory
C - Launch the currently flashed ROM

In addition, when flashing a ROM, the directory
listing box is now cleared then used as a
makeshift status bar for the erase and flash
operations. It's not pretty, but it's better than
staring at an unchanging screen for a few minutes
while hoping it is actually doing something.